Added more features.
Added scrollbars to the page image. 
Added a scrollbar to the page list.
Created the "translation" class and enabled adding objects to the page.
Enabled objects scaling when the page is zoomed.
Fixed a bug where translation objects would teleport when zooming. (Cause: Image placement coordinates were relative to the center of the image, which caused the NW-anchored canvas object coordinates to mess up when zooming changed the size of the canvas)